fix: change int to long

+++ contract/Forest.Contracts.Auction/AuctionContractConstants.cs
@@ -8,5 +8,5 @@ public static class AuctionContractConstants
     // 721 type NFT
     public const long DefaultAmount = 1;
 
-    public const int TenThousandConstant = 10000;
+    public const long TenThousandConstant = 10000;
 }
